chore: rename atoms to atomsConfig in SitecoreProvider
